move custom errors to types dir
