+++ src/jobs/ega/types/errors.ts
